+++ pytorch_segmentation_models_trainer/tools/data_handlers/raster_reader.py
@@ -24,7 +24,9 @@ from collections import OrderedDict
 from dataclasses import dataclass, field
 from enum import Enum
 from pathlib import Path
-from typing import List
+from typing import Any, Dict, List
+from affine import Affine
+import json
 
 import numpy as np
 import pandas as pd
@@ -94,7 +96,8 @@ class RasterFile:
     def build_mask(self, input_vector_layer: GeoDF,\
             output_dir: Path, output_dir_dict: OrderedDict, output_filename: str =None,\
             mask_types: List[GeomType] = None, filter_area: float = None, output_extension: str = None,\
-            compute_crossfield: bool = False, compute_distances: bool = False, compute_sizes: bool = False
+            compute_crossfield: bool = False, compute_distances: bool = False, compute_sizes: bool = False,
+            compute_bbox: bool = False
         ) -> List[str]:
         """Build segmentation mask according to parameters.
 
@@ -130,13 +133,17 @@ class RasterFile:
             compute_sizes=compute_sizes,
             compute_crossfield=compute_crossfield
         )
+        bbox_list = self.build_image_bb_annotations_from_vector_layer(
+            mask_feats, profile['transform']
+        ) if compute_bbox else None
         raster_ds.close()
         return self.write_masks_to_disk(
             raster_dict=raster_dict,
             profile=profile,
             output_dir_dict=output_dir_dict,
             output_filename=output_filename,
-            output_extension=output_extension
+            output_extension=output_extension,
+            bbox_list=bbox_list
         )
     
     def _prepare_profile(self, raster_ds, output_extension: str):
@@ -162,7 +169,7 @@ class RasterFile:
     
     def write_masks_to_disk(self, raster_dict, profile,\
             output_dir_dict, output_filename, output_extension,\
-            folder_basename=None, replicate_input_structure=True
+            folder_basename=None, replicate_input_structure=True, bbox_list=None
         ) -> List[str]:
         path_dict = OrderedDict()
         input_name, _ = os.path.basename(self.file_name).split('.')
@@ -173,6 +180,12 @@ class RasterFile:
         for key, path in output_dir_dict.items():
             if key not in raster_dict:
                 path_dict[key] = None
+                continue
+            if key == 'bounding_boxes':
+                self.save_bounding_boxes_to_disk(
+                    raster_dict['bounding_boxes'],
+                    output=os.path.join(path, subfolders, f'{output_filename}.json')
+                )
                 continue
             profile['count'] = 1 if len(raster_dict[key].shape) == 2\
                 else min(raster_dict[key].shape)
@@ -187,10 +200,14 @@ class RasterFile:
                     out.write(reshape_as_raster(raster_dict[key]))
             path_dict[key] = output
         return path_dict
+    
+    def save_bounding_boxes_to_disk(self, bounding_boxes, output: str):
+        with open(output, 'w') as out:
+            json.dump(bounding_boxes, out)
 
     def build_numpy_mask_from_vector_layer(self, mask_feats: GeoSeries, output_shape:tuple,\
         transform, mask_types: List[GeomType], compute_distances=True, compute_sizes=True,\
-        compute_crossfield=False
+        compute_crossfield=False, compute_bbox=False
     ) -> np.ndarray:
         """Builds numpy mask from vector layer using rasterio.
 
@@ -223,6 +240,10 @@ class RasterFile:
                 transform,
                 line_width=4
             )
+        if compute_bbox:
+            return_dict['bounding_boxes'] = self.build_image_bb_annotations_from_vector_layer(
+                polygons,
+                transform=transform,
         return return_dict
     
     def get_image_stats(self):
@@ -237,6 +258,28 @@ class RasterFile:
         }
         raster_ds.close()
         return return_dict
+
+    def build_image_bb_annotations_from_vector_layer(
+            self, vector_feats: GeoSeries, transform: Affine=None
+        ) -> List[Dict[str, Any]]:
+        """Builds bounding box annotations from vector layer.
+
+        Args:
+            vector_feats ([type]): [description]    
+
+        Returns:
+            [type]: [description]
+        """
+        return_list = []
+        if transform is None:
+            with rasterio.open(self.file_name) as raster_ds:
+                transform = raster_ds.transform
+        for feat in vector_feats:
+            return_list.append({
+                'class': feat.geometry.type,
+                'bbox': [~transform * point for point in np.array(feat.geometry.bounds)],
+            })
+        return return_list
 
 @dataclass
 class AbstractRasterPathListGetter(ABC):